+++ packages/cryptography/src/index.ts
@@ -1,1 +1,3 @@
+export * from './hmac.js';
 export * from './sha.js';
+export type * from './types.js';