+++ scripts/manage-codeowners.ts
@@ -1,5 +1,3 @@
-#!/usr/bin/env yarn tsx
-
 import { main } from './manage-codeowners/main.js';
 
 main().catch(function (error) {